requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
command-not-found==0.3
cryptography==3.4.8
dataclasses-json==0.6.3
dbus-python==1.2.18
debugpy==1.8.0
decorator==4.4.2
distro==1.7.0
distro-info===1.1build1
docopt==0.6.2
elevenlabs==0.2.27
ensta==5.0.3.9
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
exrex==0.11.0
Flask==3.0.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.2
httpx==0.26.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==4.6.4
inflect==7.0.0
inflection==0.5.1
insta-scrape==2.1.2
instagrapi==2.0.1
ipykernel==6.28.0
ipython==8.19.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
Js2Py==0.74
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.6.0
keyring==23.5.0
langchain==0.1.0
langchain-community==0.0.12
langchain-core==0.1.10
langsmith==0.0.80
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib-inline==0.1.6
more-itertools==8.10.0
moviepy==1.0.3
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.0
openai==1.7.2
openpyxl==3.0.10
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
pipwin==0.5.2
platformdirs==4.1.0
proglog==0.1.10
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyairtable==2.2.2
PyAudio==0.2.14
pycryptodome==3.19.0
pycryptodomex==3.18.0
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
PyGObject==3.42.1
pyjsparser==2.7.1
PyJWT==2.3.0
pyparsing==2.4.7
PyPrind==2.11.3
pySmartDL==1.3.4
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==5.4.1
pyzmq==25.1.2
requests==2.31.0
requests-futures==1.0.1
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.25
stack-data==0.6.3
stem==1.8.2
systemd-python==234
tenacity==8.2.3
tornado==6.4
torrequest==0.1.0
tqdm==4.66.1
traitlets==5.14.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.1.0
wadllib==1.3.6
wcwidth==0.2.12
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4
zipp==1.0.0
python-dotenv